# perf(rust): add MessageStream benches