--- a/rust/Cargo.toml
+++ b/rust/Cargo.toml
@@ -40,6 +40,7 @@ anyhow = "1"
 atty = "0.2"
 camino = "1.0"
 clap = { version = "3.2", features = ["derive"]}
+criterion = "0.5.1"
 itertools = "0.10"
 memmap = "0.7"
 rayon = "1.5"
@@ -47,3 +48,12 @@ serde = { version = "1.0.145", features = ["derive"] }
 serde_json = "1"
 simplelog = "0.12"
 tempfile = "3.3"
+
+[[bench]]
+name = "reader"
+harness = false
+
+[profile.bench]
+opt-level = 3
+debug = true
+lto = true