Benches MessageStream on 1M messages uncompressed/zstd/lz4.

### Baseline (m1 pro)

```
mcap_read/MessageStream_1M_uncompressed
                        time:   [67.954 ms 68.277 ms 68.946 ms]
                        thrpt:  [14.504 Melem/s 14.646 Melem/s 14.716 Melem/s]
mcap_read/MessageStream_1M_lz4
                        time:   [337.77 ms 344.16 ms 350.76 ms]
                        thrpt:  [2.8510 Melem/s 2.9056 Melem/s 2.9606 Melem/s]
mcap_read/MessageStream_1M_zstd
                        time:   [381.62 ms 384.54 ms 388.35 ms]
                        thrpt:  [2.5750 Melem/s 2.6005 Melem/s 2.6204 Melem/s]
```